feat: add logout/withdraw dialog in setting page
