+++ src/components/setting/index.ts
@@ -1,4 +1,5 @@
-export * from "./account-setting";
-export * from "./alarm-setting";
-export * from "./service-info";
+export * from "./account-setting"
+export * from "./alarm-setting"
+export * from "./service-info"
+export * from "./user-info"
